Issue 53:
Remove file-level error reporting and instead stick with record-level reporting.

Previously if even a single record contained an xml schema validation error, the whole
file would be treated as an error (skipped).  The default split size is 10,000 which means that
this occurs way to often and makes it very difficult to pinpoint and fix the actual errors.

Now we validate on a record-per-record basis, and when an error occurs we save the record
to the error_xml directory with a prefix of errors_records_in_




git-svn-id: https://xcoaitoolkit.googlecode.com/svn/trunk@72 3eaa8e94-081b-11de-9c5c-7db7966a26d7