Fix Issue #63
Previously, when / if a duplicate record was encountered in the same pass (of the same input marc file), it was unable to acquire its current OAIID value from the lucene index (because it wasn't flushed in time).  Instead, it simply skipped the record.

To fix this problem, I now flush the lucene index after each input XML record (note: not each marc file, but each split XML record which resulted from the convert process).  Also, when processing each XML record, I cache the record IDs and OAIIDs so that I may then use this information to update any duplicate or updated records (for the same ID) we may encounter during this same pass.



git-svn-id: https://xcoaitoolkit.googlecode.com/svn/trunk@90 3eaa8e94-081b-11de-9c5c-7db7966a26d7