Fix Issue #70
We now handle fatal error (marc exceptions) gracefully.

I also added an "invalid files" statistic to the modify step (similar to the import step).  I did this because I couldn't simply write out the bad records to the error_xml directory (because the marc xml libraries couldn't read the original to begin with ;-).

I also noticed that during the import process, the bad records writer wasn't being closed on the last file being processed.  This meant that if there were any error records in the last file, they would not get written (0 byte file instead).  I had to add a "cleanup" method on the IImporter interface (since the setCurrentFile(), which is the initialization phase of the reusable object, wasn't enough).




git-svn-id: https://xcoaitoolkit.googlecode.com/svn/trunk@91 3eaa8e94-081b-11de-9c5c-7db7966a26d7